move ContractDeployer to sdk.ts
